Modules split (#6)
* Split modules. Not working

* Fix module split

* Init 1.3.1 release

* Change init update description
