+++ audiobridge/db/database.py
@@ -1,8 +1,14 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-import os, psycopg2, logging, sys
+import logging
+import os, sys
+from os.path import realpath
+import psycopg2
 from psycopg2 import Error
+
+
+logger = logging.getLogger('logger')
 
 class DataBase():
 	"""Интерфейс для работы с базой данных PostgreSql.
@@ -11,7 +17,7 @@ class DataBase():
 	def __init__(self):
 		"""Инициализация класса DataBase.
 		"""
-		self.logger = logging.getLogger('logger')
+		logger = logging.getLogger('logger')
 
 		# Инициализация объекта подключения
 		self.conn = None
@@ -22,7 +28,7 @@ class DataBase():
 		"""Подключение к базе данных.
 		"""
 		try:
-			self.logger.debug(f'Connecting to {str(os.environ["DB_NAME"])}')
+			logger.debug(f'Connecting to {str(os.environ["DB_NAME"])}')
 			self.conn = psycopg2.connect(
 				user     = str(os.environ["PG_USER"]).strip(),
 				password = str(os.environ["PG_PASSWORD"]).strip(),
@@ -31,11 +37,11 @@ class DataBase():
 				database = str(os.environ["DB_NAME"]).strip())
 			self.conn.autocommit=True
 		except (Exception, Error) as er:
-			self.logger.error(f'Connection to database failed: {er}')
-			self.logger.debug('Closing the program...')
+			logger.error(f'Connection to database failed: {er}')
+			logger.debug('Closing the program...')
 			sys.exit()
 		else:
-			self.logger.debug('Database was connected successfully')
+			logger.debug('Database was connected successfully')
 			self.create_tables()
 
 	#Создание таблиц, если они отсутствуют
@@ -43,16 +49,17 @@ class DataBase():
 		"""Создание необходимых таблиц в случае их отсутствия.
 		"""
 		try:
+			path_dir = os.path.dirname(__file__)
 			with self.conn.cursor() as curs:
-				curs.execute(open("src/db/scripts/init_tables.sql", "r").read())
+				curs.execute(open(realpath(path_dir + "/scripts/init_tables.sql"), "r").read())
 
 		except (Exception, Error) as er:
 			#Закрытие освобождение памяти + выход из программы для предотвращения рекурсии и настройки PostgreSQL на хосте
-			self.logger.error(f'Tables creation failed: {er}')
+			logger.error(f'Tables creation failed: {er}')
 			if self.conn:
 				self.conn.close()
-				self.logger.debug('Connection was closed')
-			self.logger.debug('Closing the program...')
+				logger.debug('Connection was closed')
+			logger.debug('Closing the program...')
 			sys.exit()
 		else:
-			self.logger.debug(f'Tables are ready to use')
+			logger.debug(f'Tables are ready to use')
